* Changed packet bundles to support 32-bit bundles, since a test with process spy on firefox generated an initial bundle in the hundreds of thousands. * Added some todos. * Heap Id support in heap walker, and some misc readability improvements. * Minor threading fix for rendering hover/select blocks.

+++ Collector/Transport.cpp
@@ -29,7 +29,8 @@ Buffer Transport::BuildFinalBuffer(unsigned short version)
 
 	buffer.Add(&version, sizeof(version));
 
-	unsigned short numItems = PacketQueue.GetNumItems();
+	// TODO: Architecture
+	unsigned int numItems = PacketQueue.GetNumItems();
 	buffer.Add(&numItems, sizeof(numItems));
 
 	for (unsigned int i = 0; i < numItems; i++)
